# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/test/getdatabytime": {
            "get": {
                "description": "Delete user account",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete user account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    }
                }
            }
        },
        "/api/v1/user/delete": {
            "post": {
                "description": "Delete user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete user account",
                "parameters": [
                    {
                        "description": "Delete user account",
                        "name": "delete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.DeleteAccount"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "User login and get a authorization token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User login and get a authorization token",
                "parameters": [
                    {
                        "description": "User login and get a authorization token",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.UserOperation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "description": "register a new username and secret password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "register a new username and secret password",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.UserOperation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/view.ResponseMSG"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "view.DeleteAccount": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "view.ResponseMSG": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "view.UserOperation": {
            "type": "object",
            "properties": {
                "passWord": {
                    "type": "string"
                },
                "userName": {
                    "type": "object"
                }
            }
        }
    }
}